Vendor Terraform providers and modules in container image
This commit extends the Dockerfile and pipeline to vendor Terraform
providers and modules. This further decreases the time a pipeline
run takes and reduces the risk of pipeline runs failing due to
unavailability of external services.

This both improves developer experience and automation reliability.

Terraform will transparently download module and provider versions
that are not cached.